feat: refactor App component and introduce Providers for context and theme management
