+++ src/components/Providers.tsx
@@ -1,7 +1,17 @@
+import ThemeProvider from "@/components/theme";
 import ReduxProvider from "@/store/ReduxProvider";
+import LayoutContextProvider from "@/contexts/LayoutContext";
+import { ToastContainer } from 'react-toastify';
 
 const Providers = ({ children }: { children: ReactNode }) => {
-  return <ReduxProvider>{children}</ReduxProvider>;
+  return (
+    <LayoutContextProvider>
+      <ThemeProvider>
+        <ReduxProvider>{children}</ReduxProvider>
+        <ToastContainer />
+      </ThemeProvider>
+    </LayoutContextProvider>
+  );
 }
 
 export default Providers;
