+++ src/store/store.ts
@@ -1,4 +1,7 @@
+// Third-party Imports
 import { configureStore } from '@reduxjs/toolkit'
+
+// Slice Imports
 import exampleReducer from './slices/exampleSlice'
 import loginReducer from './slices/loginSlice'
