+++ src/types/base.ts
@@ -1,0 +1,6 @@
+// React Imports
+import type { ReactNode } from 'react'
+
+export type ChildrenType = {
+  children: ReactNode
+}
